+++ tests/pools/cluster/dynamic.test.js
@@ -66,7 +66,7 @@ describe('Dynamic cluster pool test suite', () => {
 
   it('Validation of inputs test', () => {
     expect(() => new DynamicClusterPool(min)).toThrowError(
-      new Error('Please specify a file with a worker implementation')
+      'Please specify a file with a worker implementation'
     )
   })