+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Fix worker function type definition and validation.
+
 ## [2.4.8] - 2023-04-12
 
 ### Fixed